Unify spatialite provider tests
And enable pyspatialite on travis
