+++ tests/src/python/CMakeLists.txt
@@ -50,7 +50,6 @@ IF (ENABLE_PGTEST)
 ENDIF (ENABLE_PGTEST)
 ADD_PYTHON_TEST(PyQgsSpatialiteProvider test_provider_spatialite.py)
 ADD_PYTHON_TEST(PyQgsShapefileProvider test_provider_shapefile.py)
-ADD_PYTHON_TEST(PyQgsSpatialiteProviderOther test_qgsspatialiteprovider.py)
 IF (WITH_APIDOC)
   ADD_PYTHON_TEST(PyQgsDocCoverage test_qgsdoccoverage.py)
 ENDIF (WITH_APIDOC)
